requirements.txt
# Core Dependencies
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
PyYAML>=6.0
pathlib2>=2.3.7

# Web Interface
streamlit>=1.25.0

# Optional GPU Support
# torch-audio>=2.0.0  # Uncomment for audio processing
# torch-vision>=0.15.0  # Uncomment for additional vision models
